+++ spec/classes/galera_debian_spec.rb
@@ -18,7 +18,7 @@ describe 'galera' do
       root_password: 'test',
       status_password: 'nonempty',
       vendor_type: 'percona',
-      vendor_version: '5.7',
+      vendor_version: '8.0',
       wsrep_group_comm_port: 4567,
       wsrep_inc_state_transfer_port: 4568,
       wsrep_sst_method: 'rsync',
@@ -142,7 +142,7 @@ describe 'galera' do
         facts
       end
 
-      case facts[:osfamily]
+      case facts[:os]['family']
       when 'Debian'
         if facts[:os]['name'] == 'Debian' && Puppet::Util::Package.versioncmp(facts[:os]['release']['major'], '12') >= 0
           it_configures 'galera on Debian'