+++ spec/classes/galera_repo_spec.rb
@@ -12,7 +12,7 @@ describe 'galera' do
   shared_examples_for 'repo on RedHat-family' do
     context 'for codership' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'codership', vendor_version: '5.7')
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0')
       end
       it { is_expected.to contain_galera__repo__config('codership') }
       it { is_expected.to contain_galera__repo__config('codership_lib') }
@@ -22,7 +22,7 @@ describe 'galera' do
 
     context 'for codership with wsrep_sst_method=xtrabackup' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'codership', vendor_version: '5.7', wsrep_sst_method: 'xtrabackup')
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0', wsrep_sst_method: 'xtrabackup')
       end
       it { is_expected.to contain_galera__repo__config('codership') }
       it { is_expected.to contain_galera__repo__config('codership_lib') }
@@ -34,7 +34,7 @@ describe 'galera' do
 
     context 'for mariadb' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.3')
+        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.11')
       end
       it { is_expected.to contain_galera__repo__config('mariadb') }
       it { is_expected.to contain_yumrepo('galera_mariadb').with(enabled: 1) }
@@ -42,7 +42,7 @@ describe 'galera' do
 
     context 'for mariadb with wsrep_sst_method=xtrabackup-v2' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.3', wsrep_sst_method: 'xtrabackup-v2')
+        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.11', wsrep_sst_method: 'xtrabackup-v2')
       end
       it { is_expected.to contain_galera__repo__config('mariadb') }
       it { is_expected.to contain_galera__repo__config('percona_tools') }
@@ -52,7 +52,7 @@ describe 'galera' do
 
     context 'for percona' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'percona', vendor_version: '5.7')
+        params.deep_merge!(vendor_type: 'percona', vendor_version: '8.0')
       end
       it { is_expected.to contain_galera__repo__config('percona') }
       it { is_expected.to contain_galera__repo__config('percona_tools') }
@@ -62,7 +62,7 @@ describe 'galera' do
 
     context 'with configure_repo=false' do
       before(:each) do
-        params.deep_merge!(configure_repo: false)
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0', configure_repo: false)
       end
       it { is_expected.not_to contain_galera__repo__config('codership') }
       it { is_expected.not_to contain_galera__repo__config('codership_lib') }
@@ -77,12 +77,15 @@ describe 'galera' do
     end
 
     context 'with epel_needed=true (default)' do
+      before(:each) do
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0')
+      end
       it { is_expected.to contain_class('epel') }
     end
 
     context 'with epel_needed=false' do
       before(:each) do
-        params.deep_merge!(epel_needed: false)
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0', epel_needed: false)
       end
       it { is_expected.not_to contain_class('epel') }
     end
@@ -91,7 +94,7 @@ describe 'galera' do
   shared_examples_for 'repo on Debian-family' do
     context 'for codership' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'codership', vendor_version: '5.7')
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0')
       end
       it { is_expected.to contain_galera__repo__config('codership') }
       it { is_expected.to contain_galera__repo__config('codership_lib') }
@@ -101,7 +104,7 @@ describe 'galera' do
 
     context 'for codership with wsrep_sst_method=xtrabackup' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'codership', vendor_version: '5.7', wsrep_sst_method: 'xtrabackup')
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0', wsrep_sst_method: 'xtrabackup')
       end
       it { is_expected.to contain_galera__repo__config('codership') }
       it { is_expected.to contain_galera__repo__config('codership_lib') }
@@ -113,14 +116,14 @@ describe 'galera' do
 
     context 'for mariadb' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.3')
+        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.11')
       end
       it { is_expected.to contain_apt__source('galera_mariadb').with(repos: 'main') }
     end
 
     context 'for mariadb with wsrep_sst_method=xtrabackup-v2' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.3', wsrep_sst_method: 'xtrabackup-v2')
+        params.deep_merge!(vendor_type: 'mariadb', vendor_version: '10.11', wsrep_sst_method: 'xtrabackup-v2')
       end
       it { is_expected.to contain_galera__repo__config('mariadb') }
       it { is_expected.to contain_galera__repo__config('percona_tools') }
@@ -130,7 +133,7 @@ describe 'galera' do
 
     context 'for percona' do
       before(:each) do
-        params.deep_merge!(vendor_type: 'percona', vendor_version: '5.7')
+        params.deep_merge!(vendor_type: 'percona', vendor_version: '8.0')
       end
       it { is_expected.to contain_galera__repo__config('percona') }
       it { is_expected.to contain_galera__repo__config('percona_tools') }
@@ -140,7 +143,7 @@ describe 'galera' do
 
     context 'with configure_repo=false' do
       before(:each) do
-        params.deep_merge!(configure_repo: false)
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0', configure_repo: false)
       end
       it { is_expected.not_to contain_galera__repo__config('codership') }
       it { is_expected.not_to contain_galera__repo__config('codership_lib') }
@@ -156,7 +159,7 @@ describe 'galera' do
 
     context 'with epel_needed=true should do nothing' do
       before(:each) do
-        params.deep_merge!(epel_needed: true)
+        params.deep_merge!(vendor_type: 'codership', vendor_version: '8.0', epel_needed: true)
       end
       it { is_expected.not_to contain_class('epel') }
     end
@@ -168,7 +171,7 @@ describe 'galera' do
         facts
       end
 
-      case facts[:osfamily]
+      case facts[:os]['family']
       when 'RedHat'
         it_configures 'repo on RedHat-family'
       when 'Debian'